# Conflicts:
#	service-order/service-order-core/src/main/java/com/wanmi/sbc/order/trade/service/TradePushERPService.java
#	service-order/service-order-core/src/main/java/com/wanmi/sbc/order/trade/service/TradeService.java

+++ service-order/service-order-core/src/main/java/com/wanmi/sbc/order/trade/model/entity/value/TradePrice.java
@@ -196,7 +196,7 @@ public class TradePrice implements Serializable {
     /**
      * 实际积分
      */
-    private Long actualPoints;
+    private BigDecimal actualPoints;
     /**
      * 实际知豆
      */
